scripts.imports-asp: Automate further
If the previous files were not deleted `unzip` would failed, it expect
only one argument but the `*` was expanded by the shell, resulting in
two (or more) files given as arguments.
Also, our documentation tell us that we should delete the current week
files once the imports are done, but to keep a backup (just in case)
until the end of the week.

So I thought that we could change the script, so it works when there is
multiple files in the directory, and so we don't have to be bothered
with the backuping, then deleting, of the current week files.

And because there is a possibility that the files don't exists (ie.
script launched too early on monday) then we check they are here.

Also make shellcheck happy for the `cd` line.

+++ scripts/imports-asp.sh
@@ -17,15 +17,27 @@ echo "Running the ASP import script"
 # Every command in this script will be written before being executed
 set +x
 
-cd $APP_HOME
+cd $APP_HOME || exit 255
+
+FLUX_IAE_FILE=$(find asp_shared_bucket/ -name 'fluxIAE_*.zip' -type f -mtime -5)
+if [[ ! -f "$FLUX_IAE_FILE" ]]; then
+    echo "Missing the flux IAE file."
+    exit 0
+fi
+
+CONTACT_EA_FILE=$(find asp_shared_bucket/ -name 'Liste_Contact_EA_*.zip' -type f -mtime -5)
+if [[ ! -f "$CONTACT_EA_FILE" ]]; then
+    echo "Missing the contact EA file."
+    exit 0
+fi
 
 # Create the "data" directory inside of the app and clear it of any previously existing data
 mkdir -p itou/siaes/management/commands/data/
 rm -rf itou/siaes/management/commands/data/*
 
 # Unzip ASP files
-unzip -P $ASP_UNZIP_PASSWORD asp_shared_bucket/fluxIAE_*.zip -d itou/siaes/management/commands/data/
-unzip -P $ASP_UNZIP_PASSWORD asp_shared_bucket/Liste_Contact_EA*.zip -d itou/siaes/management/commands/data/
+unzip -P $ASP_UNZIP_PASSWORD "$FLUX_IAE_FILE" -d itou/siaes/management/commands/data/
+unzip -P $ASP_UNZIP_PASSWORD "$CONTACT_EA_FILE" -d itou/siaes/management/commands/data/
 
 # Perform the necessary data imports
 mkdir -p $OUTPUT_PATH/populate_metabase_fluxiae
@@ -38,3 +50,6 @@ time ./manage.py import_ea_eatt --verbosity=2 |& tee -a "$OUTPUT_PATH/import_ea_
 
 # Destroy the cleartext ASP data
 rm -rf itou/siaes/management/commands/data/
+
+# Remove ASP files from 2 weeks ago
+find asp_shared_bucket/ -name '*.zip' -type f -mtime +13 -delete
